1318 - move blog folder to client
