+++ src/components/client/blog/FeaturedImage.tsx
